コンパイルエラーを修正

+++ ViewCs/Views/MainWindow.xaml.cs
@@ -45,8 +45,8 @@ public  partial class  MainWindow : Window
 
         IntPtr  hDC = grpBuffer.GetHdc();
         WinAPI.BitBlt(hDC, 8, 8, 184, 84, hDisplayDC,
-            SystemParameters.PrimaryScreenWidth - 184,
-            SystemParameters.PrimaryScreenHeight - 84,
+            (int)(SystemParameters.PrimaryScreenWidth - 184),
+            (int)(SystemParameters.PrimaryScreenHeight - 84),
             WinAPI.SRCCOPY);
         grpBuffer.ReleaseHdc(hDC);
